+++ plugins/deep-link/src/error.rs
@@ -8,8 +8,21 @@ pub type Result<T> = std::result::Result<T, Error>;
 
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
+    #[error("unsupported platform")]
+    UnsupportedPlatform,
     #[error(transparent)]
     Io(#[from] std::io::Error),
+    #[error(transparent)]
+    Tauri(#[from] tauri::Error),
+    #[cfg(target_os = "windows")]
+    #[error(transparent)]
+    Windows(#[from] windows_result::Error),
+    #[cfg(target_os = "linux")]
+    #[error(transparent)]
+    Ini(#[from] ini::Error),
+    #[cfg(target_os = "linux")]
+    #[error(transparent)]
+    ParseIni(#[from] ini::ParseError),
     #[cfg(mobile)]
     #[error(transparent)]
     PluginInvoke(#[from] tauri::plugin::mobile::PluginInvokeError),