+++ plugins/http/guest-js/index.ts
@@ -107,7 +107,7 @@ export async function fetch(
   const proxy = init?.proxy;
 
   // Remove these fields before creating the request
-  if (init != null) {
+  if (init) {
     delete init.maxRedirections;
     delete init.connectTimeout;
     delete init.proxy;
@@ -115,33 +115,47 @@ export async function fetch(
 
   const signal = init?.signal;
 
-  const headers =
-    init?.headers == null
-      ? []
-      : init.headers instanceof Headers
-        ? Array.from(init.headers.entries())
-        : Array.isArray(init.headers)
-          ? init.headers
-          : Object.entries(init.headers);
-
-  const mappedHeaders: Array<[string, string]> = headers.map(([name, val]) => [
-    name,
-    // we need to ensure we have all values as strings
-    // eslint-disable-next-line
-    typeof val === "string" ? val : (val as any).toString(),
-  ]);
+  const headers = init?.headers
+    ? init.headers instanceof Headers
+      ? init.headers
+      : new Headers(init.headers)
+    : new Headers();
 
   const req = new Request(input, init);
   const buffer = await req.arrayBuffer();
-  const reqData =
+  const data =
     buffer.byteLength !== 0 ? Array.from(new Uint8Array(buffer)) : null;
+
+  // append new headers created by the browser `Request` implementation,
+  // if not already declared by the caller of this function
+  for (const [key, value] of req.headers) {
+    if (!headers.get(key)) {
+      headers.set(key, value);
+    }
+  }
+
+  const headersArray =
+    headers instanceof Headers
+      ? Array.from(headers.entries())
+      : Array.isArray(headers)
+        ? headers
+        : Object.entries(headers);
+
+  const mappedHeaders: Array<[string, string]> = headersArray.map(
+    ([name, val]) => [
+      name,
+      // we need to ensure we have all header values as strings
+      // eslint-disable-next-line
+      typeof val === "string" ? val : (val as any).toString(),
+    ],
+  );
 
   const rid = await invoke<number>("plugin:http|fetch", {
     clientConfig: {
       method: req.method,
       url: req.url,
       headers: mappedHeaders,
-      data: reqData,
+      data,
       maxRedirections,
       connectTimeout,
       proxy,