+++ plugins/shell/build.rs
@@ -5,11 +5,26 @@
 #[path = "src/scope_entry.rs"]
 mod scope_entry;
 
-const COMMANDS: &[&str] = &["execute", "stdin_write", "kill", "open"];
+const COMMANDS: &[&str] = &["execute", "spawn", "stdin_write", "kill", "open"];
 
 fn main() {
     tauri_plugin::Builder::new(COMMANDS)
         .global_api_script_path("./api-iife.js")
         .global_scope_schema(schemars::schema_for!(scope_entry::Entry))
+        .android_path("android")
+        .ios_path("ios")
         .build();
+
+    let target_os = std::env::var("CARGO_CFG_TARGET_OS").unwrap();
+    let mobile = target_os == "ios" || target_os == "android";
+    alias("desktop", !mobile);
+    alias("mobile", mobile);
+}
+
+// creates a cfg alias if `has_feature` is true.
+// `alias` must be a snake case string.
+fn alias(alias: &str, has_feature: bool) {
+    if has_feature {
+        println!("cargo:rustc-cfg={alias}");
+    }
 }