+++ plugins/websocket/guest-js/index.ts
@@ -52,7 +52,7 @@ export default class WebSocket {
       });
     };
 
-    if (config?.headers != null) {
+    if (config?.headers) {
       config.headers = Array.from(new Headers(config.headers).entries());
     }